Format base_ts, add ability to move channel to existing channel number
Format the base_ts.ts file and change declarations of arrays and objects to use [] and {} instead of using new array and new object.

Add ability for channels to be moved to an already occupied channel number. This will shift the channel that previously occupied that channel number and all other channels after up by one number until either an empty channel number is found or the end of the channels is reached. Allows the ability to easily move channels around without the need of manually renumbering channels.