Login page fixes, add checkbox for shift channel
Fixes login page not submitting when pressing enter, also centered the login button.

Added checkbox for new shift channel feature that is on by default. Unchecking the box reverts the mapping to the old method.

Removed "(Emby only)" from kids category as this category is now supported in Plex as well.