Add TV logo lookup, general cleanup, add default TV logo
Add the ability to search Tapiosinn's tv-logo repository and easily select a TV channel logo to use from the list.

Added the ability to preview a TV channel logo URL so when selecting an image from the list or pasting one in it will show a preview of the image.

Some general linter cleanup as well as added a default TV channel logo to be used if one does not exist and added 2 scripts one for powershell and one for bash to easily generate the webUI.go file from the html folder for development.