Change-to-github-api (#19)
* Format base_ts, add ability to move channel to existing channel number

Format the base_ts.ts file and change declarations of arrays and objects to use [] and {} instead of using new array and new object.

Add ability for channels to be moved to an already occupied channel number. This will shift the channel that previously occupied that channel number and all other channels after up by one number until either an empty channel number is found or the end of the channels is reached. Allows the ability to easily move channels around without the need of manually renumbering channels.

* Login page fixes, add checkbox for shift channel

Fixes login page not submitting when pressing enter, also centered the login button.

Added checkbox for new shift channel feature that is on by default. Unchecking the box reverts the mapping to the old method.

Removed "(Emby only)" from kids category as this category is now supported in Plex as well.

* Update version number

* Add TV logo lookup, general cleanup, add default TV logo

Add the ability to search Tapiosinn's tv-logo repository and easily select a TV channel logo to use from the list.

Added the ability to preview a TV channel logo URL so when selecting an image from the list or pasting one in it will show a preview of the image.

Some general linter cleanup as well as added a default TV channel logo to be used if one does not exist and added 2 scripts one for powershell and one for bash to easily generate the webUI.go file from the html folder for development.

* update webUI builder scripts to use 2 spaces instead of tabs

* Add TV logo lookup, general cleanup, add default TV logo

Add the ability to search Tapiosinn's tv-logo repository and easily select a TV channel logo to use from the list.

Added the ability to preview a TV channel logo URL so when selecting an image from the list or pasting one in it will show a preview of the image.

Some general linter cleanup as well as added a default TV channel logo to be used if one does not exist and added 2 scripts one for powershell and one for bash to easily generate the webUI.go file from the html folder for development.

* Fix HostIP and HostName for m3u and xepg

Fixes HostIP and HostName not properly being used and set for m3u and xepg file creation.

* Fix memory buffer on Windows if cache path not on C: volume

On Windows machines if buffer to memory is active and the xTeVe cache is set to a folder on a volume other than C: it will cause a panic error and the stream not to play.

This fixes that by checking if we are on Windows and if so is the cache folder path not on the C: volume, if not create the volume it is located on in the memory virtual file system.

* Fix webserver overriding globalDomain

* Fix popup menu broken due to new preview channel logo

* Add missing port for setGlobalDomain HostIP/HostName

* Converted logos to github api for speed

* Added in File Download logic for M3U and XML files

* Added Description to Logo URL

* Start of Sort By Country

* Country Matching cleanup

* Added Logging Functions to js
Cleaned Up logo json to reduce size
Changed WS to allow more than one request
Log logos with 404 or no value
Changed bad logos to use default logo
Added in country selector in settings for logo selection
Ability to Upload M3U/XML files from desktop

* Return local file paths for M3U and XML

* Restore from configuration wizard

* Update to version 2.5.2

Co-authored-by: Christopher Norton <24207721+cnorton-webdev@users.noreply.github.com>